+++ packages-tooling/parcel-transformer/rollup.config.js
@@ -1,0 +1,4 @@
+import { getRollupConfig } from '../rollup-utils.js';
+import pkg from './package.json';
+
+export default getRollupConfig(pkg);