+++ packages/testing/src/startup.ts
@@ -17,10 +17,8 @@ export const onFixtureCreated = <T>(callback: (fixture: IFixture<T>) => unknown)
   });
 };
 
-export function createFixture<
-  T,
-  K = (T extends Constructable<infer U> ? U : T)
->(template: string | Node,
+export function createFixture<T, K = (T extends Constructable<infer U> ? U : T)>(
+  template: string | Node,
   $class?: T,
   registrations: unknown[] = [],
   autoStart: boolean = true,
@@ -44,7 +42,7 @@ export function createFixture<
 
   if (container.has(App, true)) {
     throw new Error(
-      'Container of the context cotains instance of the application root component. ' +
+      'Container of the context contains instance of the application root component. ' +
       'Consider using a different class, or context as it will likely cause surprises in tests.'
     );
   }
@@ -58,7 +56,7 @@ export function createFixture<
 
   let tornCount = 0;
 
-  const getBy = (selector: string): HTMLElement => {
+  const getBy: Document['querySelector'] = (selector: string): HTMLElement => {
     const elements = host.querySelectorAll<HTMLElement>(selector);
     if (elements.length > 1) {
       throw new Error(`There is more than 1 element with selector "${selector}": ${elements.length} found`);
@@ -78,12 +76,27 @@ export function createFixture<
     }
     return elements.length === 0 ? null : elements[0];
   };
-  const assertText = (selector: string, text: string): void => {
-    const el = queryBy(selector);
-    if (el === null) {
-      throw new Error(`No element found for selector "${selector}" to compare text content with "${text}"`);
+  const assertText = (selector: string, text?: string): void => {
+    if (arguments.length === 2) {
+      const el = queryBy(selector);
+      if (el === null) {
+        throw new Error(`No element found for selector "${selector}" to compare text content with "${text}"`);
+      }
+      assert.strictEqual(el.textContent, text);
+    } else {
+      assert.strictEqual(host.textContent, selector);
     }
-    assert.strictEqual(el.textContent, text);
+  };
+  const assertHtml = (selector: string, html?: string) => {
+    if (arguments.length === 2) {
+      const el = queryBy(selector);
+      if (el === null) {
+        throw new Error(`No element found for selector "${selector}" to compare innerHTML with "${html}"`);
+      }
+      assert.strictEqual(el.innerHTML, html);
+    } else {
+      assert.strictEqual(host.innerHTML, selector);
+    }
   };
   const trigger = ((selector: string, event: string, init?: CustomEventInit): void => {
     const el = queryBy(selector);
@@ -92,7 +105,7 @@ export function createFixture<
     }
     el.dispatchEvent(new ctx.CustomEvent(event, init));
   }) as ITrigger;
-  ['click', 'change', 'input'].forEach(event => {
+  ['click', 'change', 'input', 'scroll'].forEach(event => {
     Object.defineProperty(trigger, event, { configurable: true, writable: true, value: (selector: string, init?: CustomEventInit): void => {
       const el = queryBy(selector);
       if (el === null) {
@@ -101,6 +114,15 @@ export function createFixture<
       el.dispatchEvent(new ctx.CustomEvent(event, init));
     } });
   });
+
+  const scrollBy = (selector: string, init: number | ScrollToOptions) => {
+    const el = queryBy(selector);
+    if (el === null) {
+      throw new Error(`No element found for selector "${selector}" to scroll by "${JSON.stringify(init)}"`);
+    }
+    el.scrollBy(typeof init === 'number' ? { top: init } : init);
+    el.dispatchEvent(new Event('scroll'));
+  };
 
   const fixture = new class Results implements IFixture<K> {
     public startPromise = startPromise;
@@ -118,14 +140,20 @@ export function createFixture<
       await au.app({ host: host, component }).start();
     }
 
-    public async tearDown() {
+    public tearDown() {
       if (++tornCount === 2) {
         console.log('(!) Fixture has already been torn down');
         return;
       }
-      await au.stop();
-      root.remove();
-      au.dispose();
+      const dispose = () => {
+        root.remove();
+        au.dispose();
+      };
+      const ret = au.stop();
+      if (ret instanceof Promise)
+        return ret.then(dispose);
+      else
+        return dispose();
     }
 
     public get torn() {
@@ -135,15 +163,20 @@ export function createFixture<
     /**
      * @returns a promise that resolves after the associated app has started
      */
-    public get promise(): Promise<IFixture<K>> {
-      return Promise.resolve(startPromise).then(() => this as IFixture<K>);
+    public get started(): Promise<IFixture<K>> {
+      if (startPromise instanceof Promise) {
+        return Promise.resolve(startPromise).then(() => this as IFixture<K>);
+      }
+      return Promise.resolve(this);
     }
 
     public getBy = getBy;
     public getAllBy = getAllBy;
     public queryBy = queryBy;
     public assertText = assertText;
+    public assertHtml = assertHtml;
     public trigger = trigger;
+    public scrollBy = scrollBy;
   }();
 
   fixtureHooks.publish('fixture:created', fixture);
@@ -164,17 +197,127 @@ export interface IFixture<T> {
   readonly observerLocator: IObserverLocator;
   readonly torn: boolean;
   start(): Promise<void>;
-  tearDown(): Promise<void>;
-  readonly promise: Promise<IFixture<T>>;
-  getBy(selector: string): HTMLElement;
-  getAllBy(selector: string): HTMLElement[];
-  queryBy(selector: string): HTMLElement | null;
+  tearDown(): void | Promise<void>;
+  readonly started: Promise<IFixture<T>>;
+
+  /**
+   * Returns the first element that is a descendant of node that matches selectors, and throw if there is more than one, or none found
+   */
+  getBy<K extends keyof HTMLElementTagNameMap>(selectors: K): HTMLElementTagNameMap[K];
+  getBy<K extends keyof SVGElementTagNameMap>(selectors: K): SVGElementTagNameMap[K];
+  getBy<E extends HTMLElement = HTMLElement>(selectors: string): E | null;
+  /**
+   * Returns all element descendants of node that match selectors.
+   */
+  getAllBy<K extends keyof HTMLElementTagNameMap>(selectors: K): HTMLElementTagNameMap[K][];
+  getAllBy<K extends keyof SVGElementTagNameMap>(selectors: K): SVGElementTagNameMap[K][];
+  getAllBy<E extends HTMLElement = HTMLElement>(selectors: string): E[];
+  /**
+   * Returns the first element that is a descendant of node that matches selectors, and null if none found
+   */
+  queryBy<K extends keyof HTMLElementTagNameMap>(selectors: K): HTMLElementTagNameMap[K] | null;
+  queryBy<K extends keyof SVGElementTagNameMap>(selectors: K): SVGElementTagNameMap[K] | null;
+  queryBy<E extends HTMLElement = HTMLElement>(selectors: string): E | null;
+
+  /**
+   * Assert the text content of the current application host equals to a given string
+   */
+  assertText(text: string): void;
+  /**
+   * Assert the text content of an element matching the given selector inside the application host equals to a given string.
+   *
+   * Will throw if there' more than one elements with matching selector
+   */
   assertText(selector: string, text: string): void;
+
+  /**
+   * Assert the inner html of the current application host equals to the given html string
+   */
+  assertHtml(html: string): void;
+  /**
+   * Assert the inner html of an element matching the selector inside the current application host equals to the given html string.
+   *
+   * Will throw if there' more than one elements with matching selector
+   */
+  assertHtml(selector: string, html: string): void;
+
   trigger: ITrigger;
+
+  /**
+   * A helper to scroll and trigger a scroll even on an element matching the given selector
+   */
+  scrollBy(selector: string, options: number | ScrollToOptions): void;
 }
 
-export type ITrigger = ((selector: string, event: string, init: CustomEventInit) => void) & {
+export type ITrigger = ((selector: string, event: string, init?: CustomEventInit) => void) & {
   click(selector: string, init?: CustomEventInit): void;
   change(selector: string, init?: CustomEventInit): void;
   input(selector: string, init?: CustomEventInit): void;
+  scroll(selector: string, init?: CustomEventInit): void;
 };
+
+export interface IFixtureBuilderBase<T, E = {}> {
+  html(html: string): this & E;
+  html<M>(html: TemplateStringsArray, ...values: TemplateValues<M>[]): this & E;
+  component(comp: T): this & E;
+  deps(...args: unknown[]): this & E;
+}
+
+type BuilderMethodNames = 'html' | 'component' | 'deps';
+type CreateBuilder<T, Availables extends BuilderMethodNames = BuilderMethodNames> = {
+  [key in Availables]:
+    key extends 'html'
+      ? {
+        (html: string): CreateBuilder<T, Exclude<Availables, 'html'>> & { build(): IFixture<T> };
+        (html: TemplateStringsArray, ...values: TemplateValues<T>[]): CreateBuilder<T, Exclude<Availables, 'html'>> & { build(): IFixture<T> };
+      }
+        : (...args: Parameters<IFixtureBuilderBase<T>[key]>) => CreateBuilder<T, Exclude<Availables, key>>
+} & (never extends Availables ? { build(): IFixture<T> } : {});
+
+type TaggedTemplateLambda<M> = (vm: M) => unknown;
+type TemplateValues<M> = string | number | TaggedTemplateLambda<M>;
+
+class FixtureBuilder<T> {
+  private _html?: string | TemplateStringsArray;
+  private _htmlArgs?: TemplateValues<T>[];
+  private _comp?: T;
+  private _args?: unknown[];
+
+  public html(html: string | TemplateStringsArray, ...htmlArgs: TemplateValues<T>[]): CreateBuilder<T, Exclude<BuilderMethodNames, 'html'>> {
+    this._html = html;
+    this._htmlArgs = htmlArgs;
+    return this;
+  }
+
+  public component(comp: T): CreateBuilder<T, Exclude<BuilderMethodNames, 'component'>> {
+    this._comp = comp;
+    return this;
+  }
+
+  public deps(...args: unknown[]): CreateBuilder<T, Exclude<BuilderMethodNames, 'deps'>> {
+    this._args = args;
+    return this;
+  }
+
+  public build() {
+    if (this._html === void 0) {
+      throw new Error('Builder is not ready, missing template, call .html()/.html`` first');
+    }
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return createFixture<any>(
+      typeof this._html === 'string' ? this._html : brokenProcessFastTemplate(this._html, ...this._htmlArgs ?? []),
+      this._comp,
+      this._args
+    );
+  }
+}
+
+function brokenProcessFastTemplate(html: TemplateStringsArray, ..._args: unknown[]): string {
+  return html.join('');
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+createFixture.html = <T = Record<string, any>>(html: string | TemplateStringsArray, ...values: TemplateValues<T>[]) => new FixtureBuilder<T>().html(html, ...values) ;
+createFixture.component = <T>(component: T) => new FixtureBuilder<T>().component(component);
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+createFixture.deps = <T = Record<string, any>>(...deps: unknown[]) => new FixtureBuilder<T>().deps(...deps);